+++ tests/posix/poller/enc/CMakeLists.txt
@@ -3,7 +3,8 @@
 
 oeedl_file(../poller.edl enclave gen)
 
-add_executable(poller_enc enc.cpp ../client.cpp ../server.cpp ${gen})
+add_executable(poller_enc
+    enc.cpp ../client.cpp ../server.cpp ../poller.cpp ${gen})
 
 maybe_build_using_clangw(poller_enc)
