+++ tests/posix/poller/host/CMakeLists.txt
@@ -4,7 +4,8 @@
 
 oeedl_file(../poller.edl host gen)
 
-add_executable(poller_host host.cpp ../client.cpp ../server.cpp ${gen})
+add_executable(poller_host
+    host.cpp ../client.cpp ../server.cpp ../poller.cpp ${gen})
 
 target_include_directories(poller_host PRIVATE ${CMAKE_CURRENT_BINARY_DIR})
